% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_stay.R
\name{parse_video_datetime}
\alias{parse_video_datetime}
\title{Parse the recording date-time from a video file name}
\usage{
parse_video_datetime(x, tz = "UTC")
}
\arguments{
\item{x}{Character vector of file names.}

\item{tz}{Time zone of the returned date-times. Default \code{"UTC"}.}
}
\value{
A \code{POSIXct} vector; \code{NA} where the name does not match.
}
\description{
Extracts the date-time from file names of the form
\verb{<camera>_yymmdd_HHMMSS_...} (for example
\code{c15_240501_105240_05010001.MOV} gives 2024-05-01 10:52:40).
}
\examples{
parse_video_datetime("c15_240501_105240_05010001.MOV")
}
